# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Seletivo Seplag Gov-MT",
        "version": "0.1"
    },
    "paths": {
...
        "/api/servidor-efetivo": {
            "get": {
                "tags": [
                    "Servidores Efetivos"
                ],
                "summary": "Servidores Efetivos",
                "description": "Lista todos os Servidores Efetivos ligados a sua Unidade-Lotação",
                "operationId": "4b1c83d6265f7aad3977994940ab58c8",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Servidores efetivos não encontrados"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Servidores Efetivos"
                ],
                "summary": "Cadastra um Novo Servidor Efetivo",
                "description": "Registrar a Pessoa (pes_id) e seua matrícula",
                "operationId": "92a0d07a67d0101af814c73716c946be",
                "parameters": [
                    {
                        "name": "pes_id",
                        "in": "query",
                        "description": "Nº de identifição da Pessoa",
                        "required": true
                    },
                    {
                        "name": "se_matricula",
                        "in": "query",
                        "description": "Nº da matrícula",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Erro no cadastro"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/servidor-efetivo/{pes_id}": {
            "get": {
                "tags": [
                    "Servidores Efetivos"
                ],
                "summary": "Mostra um Servidor Efetivo",
                "description": "Pesquisa um Servidor Efetivo através da Pessoa (pes_id)",
                "operationId": "802fd070a68a2355208902a2c3100873",
                "parameters": [
                    {
                        "name": "pes_id",
                        "in": "path",
                        "description": "Nº de identificação da Pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pessoa Encontrada",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Pessoa não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Servidores Efetivos"
                ],
                "summary": "Atualizar dados de um Servidor Efetivo",
                "description": "Editar a matrícula de um Servidor Efetivo",
                "operationId": "dc1e4185482e3c5538db91f7b5d53e0f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pes_id",
                                    "se_matricula"
                                ],
                                "properties": {
                                    "pes_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "se_matricula": {
                                        "type": "string",
                                        "example": "00001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Servidor Efetivo atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Servidor Efetivo atualizado com sucesso"
                                        },
                                        "servidor-efetivo": {
                                            "properties": {
                                                "pes_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "se_matricula": {
                                                    "type": "string",
                                                    "example": "00001"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição inválida"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Servidores Efetivos"
                ],
                "summary": "Exclui a informação do Servidor Efetivo",
                "description": "Exclui a informação do Servidor Efetivo (pes_id)",
                "operationId": "18f59ad1bbf31bb96c8ddd22c3c6de9a",
                "parameters": [
                    {
                        "name": "pes_id",
                        "in": "path",
                        "description": "ID da Pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informação excluído com sucesso",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Não foi possível excluir a informação de servidor efetivo."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}